+++ controle_estoque/Crud/CrudRelVenda.py
@@ -2,16 +2,15 @@
 
 import peewee
 
-from orm.Conexao import Conexao, RelacaoCompra, Produto
+from Crud.Conexao import Conexao, RelacaoVenda, Produto
 
 
-class CrudRelCompra(object):
-    def __init__(self, id="", idCompra="", idProduto="", produto="",
-                 qtde="",
+class CrudRelVenda(object):
+    def __init__(self, id="", idVenda="", idProduto="", produto="", qtde="",
                  valorUnitario="", valorTotal="", obs="", query=""):
 
         self.id = id
-        self.idCompra = idCompra
+        self.idVenda = idVenda
         self.idProduto = idProduto
         self.produto = produto
         self.qtde = qtde
@@ -26,9 +25,9 @@ class CrudRelCompra(object):
 
         try:
             # Query
-            row = RelacaoCompra.insert(
+            row = RelacaoVenda.insert(
                 id=self.id,
-                id_compra=self.idCompra,
+                id_venda=self.idVenda,
                 id_produto=self.idProduto,
                 qtde=self.qtde,
                 valor_unitario=self.valorUnitario,
@@ -52,14 +51,14 @@ class CrudRelCompra(object):
         try:
 
             # Query
-            self.query = (RelacaoCompra.select(
-                RelacaoCompra, Produto.id, Produto.produto)
+            self.query = (RelacaoVenda.select(
+                RelacaoVenda, Produto.id, Produto.produto)
                 .join(Produto)
-                .where(RelacaoCompra.id_compra == self.idCompra))
+                .where(RelacaoVenda.id_venda == self.idVenda))
 
             # Convertendo variaveis em lista
             self.id = []
-            self.idCompra = []
+            self.idVenda = []
             self.idProduto = []
             self.produto = []
             self.qtde = []
@@ -70,7 +69,7 @@ class CrudRelCompra(object):
             # Salvando resultado da query e suas listas
             for row in self.query:
                 self.id.append(row.id)
-                self.idCompra.append(row.id_compra)
+                self.idVenda.append(row.id_venda)
                 self.idProduto.append(row.id_produto.id)
                 self.produto.append(row.id_produto.produto)
                 self.qtde.append(row.qtde)
@@ -91,8 +90,8 @@ class CrudRelCompra(object):
         try:
 
             # Query
-            self.query = (RelacaoCompra.delete()
-                          .where(RelacaoCompra.id == self.id))
+            self.query = (RelacaoVenda.delete()
+                          .where(RelacaoVenda.id == self.id))
 
             # Executando a query
             self.query.execute()
