+++ controle_estoque/Funcoes/comercial.py
@@ -3,7 +3,7 @@ from PySide2.QtWidgets import QCompleter, QLineEdit
 from PySide2.QtCore import QDate, QDateTime, Qt, QStringListModel
 from PySide2.QtGui import QIntValidator, QDoubleValidator
 from functools import partial
-from orm.CrudProduto import CrudProduto
+from Crud.CrudProduto import CrudProduto
 
 
 class Comercial(object):
