+++ controle_estoque/mainclientes.py
@@ -9,8 +9,8 @@ from PySide2.QtWebEngineWidgets import QWebEngineView
 
 from Views.mainClientes import Ui_ct_MainClientes
 from Views.formClientes import Ui_ct_FormClientes
-from orm.CrudCliente import CrudCliente
-from orm.CrudVenda import CrudVenda
+from Crud.CrudCliente import CrudCliente
+from Crud.CrudVenda import CrudVenda
 
 
 class MainClientes(Ui_ct_MainClientes, Ui_ct_FormClientes):
